Added additional helpers. Moved helper functions from HallwayGenerator to specific helpers
