+++ Helpers/DistanceHelper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8bb07e7f1c1ad9947aecd158ec46bcf5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
